// src/libs/irc.d.ts
    export interface IClientOpts {
        /**
         * IRC username
         * @default 'nodebot'
         */
        userName?: string;

        /**
         * IRC username
         * @default ''
         */
        password?: string;

        /**
         * IRC "real name"
         * @default 'nodeJS IRC client'
         */
        realName?: string;

        /**
         * IRC connection port. See
         * https://nodejs.org/api/net.html#net_socket_remoteport
         * @default 6667
         */
        port?: number;

        /**
         * Local interface to bind to for network connections. See
         * https://nodejs.org/api/net.html#net_socket_localaddress
         */
        localAddress?: string;

        /**
         * Should we output debug messages to STDOUT?
         * @default false
         */
        debug?: boolean;

        /**
         * Should we output IRC errors?
         * @default false
         */
        showErrors?: boolean;

        /**
         * Should we auto-rejoin channels?
         * @default false
         */
        autoRejoin?: boolean;

        /**
         * Should we auto-reconnect to networks?
         * @default true
         */
        autoConnect?: boolean;

        /**
         * Channels to join
         * @default []
         */
        channels?: string[];

        /**
         * Should SSL be used? Can either be true or crypto credentials.
         * @default false
         */
        secure?: boolean | tls.SecureContext;

        /**
         * Should we accept self-signed certificates?
         * @default false
         */
        selfSigned?: boolean;


        /**
         * Should we accept expired certificates?
         * @default false
         */
        certExpired?: boolean;

        /**
         * Should we queue our messages to ensure we don't get kicked?
         * @default false
         */
        floodProtection?: boolean;

        /**
         * Delay between messages when flood protection is active
         * @default 1000
         */
        floodProtectionDelay?: number;

        /**
         * Should we use SASL authentication?
         * @default false
         */
        sasl?: boolean;

        /**
         * Should we strip mIRC colors from the output messages?
         * @default false
         */
        stripColors?: boolean;

        /**
         * Channel prefix
         * @default '&#'
         */
        channelPrefixes?: string;

        /**
         * Characters to split a message at.
         * @default 512
         */
        messageSplit?: number;

        /**
         * Encoding to use. See
         * https://nodejs.org/api/stream.html#stream_readable_setencoding_encoding
         * @default 'utf-8'
         */
        encoding?: string;
    }

    /** Command types */
    export enum CommandType {
        normal, reply, error
    }

    /** Parsed IRC message. */
    export interface IMessage {
        /** Prefix */
        prefix?: string;

        /** Mapped IRC command */
        command: string;

        /** Raw IRC command */
        rawCommand: string;

        /** Command type */
        commandType: CommandType;

        /** Command arguments */
        args: string[];
    }